move hs download method to hs_utils and delete unnecessary methods in hs_utils.py
